commit test

+++ src/main/java/smvc/controller/CustomerController.java
@@ -44,7 +44,7 @@ public class CustomerController {
 			orderby = "asc";
 		}
 
-		int pageid = 1;
+		float pageid = 1;
 		int total = 4;
 
 		int totalRecords = customerDao.getCount();
